refactor: decouple experience capability repository
