+++ src/apps/server/experiences/providers/experience.provider.ts
@@ -4,7 +4,8 @@ import { CapabilityRepository } from '@libs/modules/database/repositories/capabi
 import { CapabilityRepositoryImpl } from '@libs/modules/database/repositories/capability/capability.repository';
 import { ExperienceRepository } from '@libs/modules/database/repositories/experience/experience.interface';
 import { ExperienceRepositoryImpl } from '@libs/modules/database/repositories/experience/experience.repository';
-import { ExperienceCapabilityRepository } from '@libs/modules/database/repositories/experienceCapability.repository';
+import { ExperienceCapabilityRepository } from '@libs/modules/database/repositories/experienceCapability/experienceCapability.interface';
+import { ExperienceCapabilityRepositoryImpl } from '@libs/modules/database/repositories/experienceCapability/experienceCapability.repository';
 import { ExperienceInfoRepository } from '@libs/modules/database/repositories/experienceInfo.repository';
 
 export const experienceProviders = [
@@ -17,7 +18,10 @@ export const experienceProviders = [
     provide: CapabilityRepository,
     useClass: CapabilityRepositoryImpl,
   },
-  ExperienceCapabilityRepository,
+  {
+    provide: ExperienceCapabilityRepository,
+    useClass: ExperienceCapabilityRepositoryImpl,
+  },
   {
     provide: AiResumeRepository,
     useClass: AiResumeRepositoryImpl,
