+++ src/libs/modules/database/repositories/experienceCapability/experienceCapability.interface.ts
@@ -1,0 +1,9 @@
+import { Prisma, ExperienceCapability } from '@prisma/client';
+
+export interface ExperienceCapabilityRepository {
+  createMany(createdInfos: { capabilityId: number; experienceId: number }[]): Promise<Prisma.BatchPayload>;
+  deleteByExperienceId(experienceId: number): Promise<Prisma.BatchPayload>;
+  findManyByFilter(where: Prisma.ExperienceCapabilityWhereInput): Promise<ExperienceCapability[]>;
+}
+
+export const ExperienceCapabilityRepository = Symbol('ExperienceCapabilityRepository');
